Rename Configuration namespace, fix data attribute lookup
